# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: Ramesh Sunkara
  description: A sample service to demonstrate how to develop REST API in golang
  title: GO Rest Example API Service (Purchase Order Tracker)
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Provides the current status/health of the service. It checks the
        health of all the critical components of the service.
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: Something is wrong
          schema:
            type: string
      summary: Status/Health of the service (Doesn't work in Swagger UI)
      tags:
      - Status
  /orders/:
    get:
      consumes:
      - application/json
      description: Fetches all orders
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Fetch all orders
      tags:
      - Fetch
    post:
      consumes:
      - application/json
      description: Used to either create or update an order
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Creates or Updates an order
      tags:
      - Fetch
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete single Order document identified by give id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: bad request
          schema:
            type: string
      summary: Delete single Order document identified by give id
      tags:
      - Fetch
    get:
      consumes:
      - application/json
      description: Fetch single Order document identified by give id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: bad request
          schema:
            type: string
      summary: Fetch single Order document identified by give id
      tags:
      - Fetch
swagger: "2.0"